% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarGraph.R
\name{BarGraph}
\alias{BarGraph}
\title{Provides a pathological association between a gene and different tissues using a bar graph}
\usage{
BarGraph(GeneName, Anatomy)
}
\arguments{
\item{GeneName}{}

\item{Anatomy}{Anatomy can be either Male or Female}
}
\description{
Provides a pathological association between a gene and different tissues using a bar graph
}
