% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biopsy.R
\name{Biopsy}
\alias{Biopsy}
\title{Provides a visual pathological association between a gene and the input tissue}
\usage{
Biopsy(GeneName, Tissue)
}
\arguments{
\item{GeneName}{}

\item{Tissue}{Name of a tissue}
}
\description{
Provides a visual pathological association between a gene and the input tissue
}
