% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionMap_PathoMap.R
\name{ExpressionMap_PathoMap}
\alias{ExpressionMap_PathoMap}
\title{Provides a visual pathological association between a gene and different tissues}
\usage{
ExpressionMap_PathoMap(GeneName, Anatomy)
}
\arguments{
\item{GeneName}{}

\item{Anatomy}{Anatomy can be either Male or Female}
}
\description{
Provides a visual pathological association between a gene and different tissues
}
